requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
langchain==0.0.335
langchain-openai==0.0.2
chromadb==0.4.18
tiktoken==0.5.2
python-multipart==0.0.6
